Panic on missing translation for custom return types
When returning a `String` we can return a nice placeholder string
if a translation is missing. But if the user specified a return type,
we cannot just return a dummy value. Thus we have to panic.

We may want to let the user specify default values for specific
return types. But that's not that important right now.